use logger and handle not fount exception

+++ src/events/events.controller.ts
@@ -8,6 +8,8 @@ import {
   Body,
   HttpCode,
   ParseIntPipe,
+  Logger,
+  NotFoundException,
 } from '@nestjs/common'
 import {
   Like,
@@ -21,6 +23,8 @@ import { Event } from './events.entity'
 
 @Controller({ path: '/events' })
 export class EventsController {
+  private readonly logger = new Logger(EventsController.name)
+
   constructor(
     @InjectRepository(Event)
     private readonly repository: Repository<Event>
@@ -28,7 +32,11 @@ export class EventsController {
 
   @Get()
   async findAll() {
-    return await this.repository.find()
+    this.logger.log('Hit the findAll rout')
+    const events = await this.repository.find()
+    this.logger.debug(`Found ${events.length} events`)
+
+    return events
   }
 
   @Get('/practice_find_query')
@@ -49,7 +57,13 @@ export class EventsController {
 
   @Get(':id')
   async findOne(@Param('id', ParseIntPipe) id: number) {
-    return await this.repository.findOne(id)
+    const event = await this.repository.findOne(id)
+
+    if (!event) {
+      throw new NotFoundException()
+    }
+
+    return event
   }
 
   @Post()
@@ -65,6 +79,10 @@ export class EventsController {
   async update(@Param('id') id, @Body() input: UpdateEventDto) {
     const event = await this.repository.findOne(id)
 
+    if (!event) {
+      throw new NotFoundException()
+    }
+
     return await this.repository.save({
       ...event,
       ...input,
@@ -78,6 +96,10 @@ export class EventsController {
   @HttpCode(204)
   async remove(@Param('id') id) {
     const event = await this.repository.findOne(id)
+
+    if (!event) {
+      throw new NotFoundException()
+    }
 
     await this.repository.remove(event)
   }
